Move service config to its own file
