+++ src/remotePipResolver.js
@@ -1,23 +1,11 @@
 'use strict';
 
-const url = require('url');
 const logger = require('pelias-logger').get('wof-admin-lookup');
 const _ = require('lodash');
 const request = require('request');
 
-const ServiceConfiguration = require('pelias-microservice-wrapper').ServiceConfiguration;
 const service = require('pelias-microservice-wrapper').service;
-
-class PointInPolygon extends ServiceConfiguration {
-  constructor(o) {
-    super('pip', o);
-  }
-
-  getUrl(params) {
-    // use resolve to eliminate possibility of duplicate /'s in URL
-    return url.resolve(this.baseUrl, `${params.lon}/${params.lat}`);
-  }
-}
+const PointInPolygon = require('./service/PointInPolygon');
 
 /**
  * RemotePIPService class
@@ -38,8 +26,7 @@ RemotePIPService.prototype.lookup = function lookup(centroid, _, callback) {
 
   this.pipService(centroid, (err, response, results) => {
     if (err) {
-      console.log(err);
-      return callback(err.message);
+      return callback(err);
     }
 
     callback(null, response);
